Leverage 'is-multi-assay' result from soft assay response to set 'assay_modality' and 'multi_assay_category'
